Rename components so they don't get packaged by webpack.
Also explicitly disable some unused plugins in the build for certain
modules; I haven't tested if this speeds up reload time but it reduces
the unnecessary output so maybe...
