--- modules/admin/app/assets/js/datamanager/components/_comp.vue
@@ -1,9 +1,0 @@
-<script>
-export default {
-
-};
-</script>
-
-<template>
-</template>
-
