+++ modules/admin/app/assets/js/datamanager/components/_delete-button.vue
