+++ modules/admin/app/assets/js/datamanager/components/_drag-handle.vue
