+++ modules/admin/app/assets/js/datamanager/components/_file-picker-suggestion.vue
