+++ modules/admin/app/assets/js/datamanager/components/_files-table.vue
