+++ modules/admin/app/assets/js/datamanager/components/_ingest-options-panel.vue
@@ -1,6 +1,6 @@
 <script>
 
-import ModalWindow from './modal-window';
+import ModalWindow from './_modal-window';
 import DAO from '../dao';
 
 import _pick from 'lodash/pick';
