+++ modules/admin/app/assets/js/datamanager/components/_log-window.vue
