+++ modules/admin/app/assets/js/datamanager/components/_mixin-preview.vue
@@ -1,4 +1,5 @@
 <script>
+
 export default {
   data: function() {
     return {
