+++ modules/admin/app/assets/js/datamanager/components/_mixin-stage.vue
