+++ modules/admin/app/assets/js/datamanager/components/_mixin-two-panel.vue
