+++ modules/admin/app/assets/js/datamanager/components/_modal-window.vue
