+++ modules/admin/app/assets/js/datamanager/components/_transformation-item.vue
