+++ modules/admin/app/assets/js/datamanager/components/_upload-progress.vue
