+++ modules/admin/app/assets/js/datamanager/components/_xquery-editor.vue
