+++ modules/admin/app/assets/js/datamanager/components/_xslt-editor.vue
