+++ package.json
@@ -32,12 +32,18 @@
   "author": "EHRI",
   "license": "EUPL-1.2",
   "devDependencies": {
+    "@types/codemirror": "^0.0.108",
     "@vue/test-utils": "^1.1.3",
+    "css-loader": "^5.1.2",
+    "fork-ts-checker-webpack-plugin": "^6.1.1",
     "grunt": "^1.3.0",
     "grunt-contrib-copy": "^1.0.0",
     "jest": "^26.6.3",
     "jsdom": "^16.4.0",
     "jsdom-global": "^3.0.2",
+    "style-loader": "^2.0.0",
+    "ts-loader": "^8.0.18",
+    "typescript": "^4.3.0-dev.20210312",
     "vue-loader": "^15.9.6",
     "vue-template-compiler": "^2.6.12",
     "webpack": "^5.24.3",
